move checker to an extension
