+++ doc/data/messages/b/bad-dunder-name/pylintrc
@@ -1,0 +1,2 @@
+[MAIN]
+load-plugins=pylint.extensions.dunder
