+++ doc/whatsnew/fragments/3038.new_check
@@ -1,3 +1,3 @@
-Added ``bad-dunder-name`` check which flags bad or misspelled dunder methods.
+Added ``bad-dunder-name`` extension check which flags bad or misspelled dunder methods.
 
 Closes #3038
