+++ tests/functional/a/arguments_differ_issue5371.py
@@ -1,4 +1,3 @@
-# pylint: disable=bad-dunder-name
 """https://github.com/PyCQA/pylint/issues/5371"""
 from enum import Enum
