+++ tests/functional/ext/bad_dunder/bad_dunder_name.rc
@@ -1,0 +1,2 @@
+[MAIN]
+load-plugins=pylint.extensions.dunder
