+++ tests/functional/ext/no_self_use/no_self_use.py
@@ -1,4 +1,4 @@
-# pylint: disable=too-few-public-methods,missing-docstring,invalid-name,bad-dunder-name
+# pylint: disable=too-few-public-methods,missing-docstring,invalid-name
 """test detection of method which could be a function"""
 from abc import ABC, abstractmethod
 from typing import Protocol, overload
