+++ tests/functional/n/name/name_styles.py
@@ -1,6 +1,6 @@
 """Test for the invalid-name warning."""
 # pylint: disable=unnecessary-pass, unnecessary-comprehension, unused-private-member
-# pylint: disable=unnecessary-lambda-assignment, bad-dunder-name
+# pylint: disable=unnecessary-lambda-assignment
 import abc
 import collections
 import typing
