+++ tests/functional/s/slots_checks.py
@@ -1,6 +1,6 @@
 """ Checks that classes uses valid __slots__ """
 
-# pylint: disable=too-few-public-methods, missing-docstring, bad-dunder-name
+# pylint: disable=too-few-public-methods, missing-docstring
 # pylint: disable=using-constant-test, wrong-import-position, no-else-return, line-too-long, unused-private-member
 from collections import deque
