+++ tests/functional/u/unexpected_special_method_signature.py
@@ -1,7 +1,7 @@
 """Test for special methods implemented incorrectly."""
 
 # pylint: disable=missing-docstring, unused-argument, too-few-public-methods
-# pylint: disable=invalid-name,too-many-arguments,bad-staticmethod-argument,bad-dunder-name
+# pylint: disable=invalid-name,too-many-arguments,bad-staticmethod-argument
 
 class Invalid:
